sensor.py renamed to inputDevices.py
